+++ lib/error.dart
@@ -1,0 +1,2 @@
+export 'src/error/generic_error.dart';
+export 'src/error/template_error.dart';