+++ lib/type_utils.dart
@@ -1,3 +1,5 @@
 export 'implementer.dart';
 export 'type_detection.dart';
 export 'builder.dart';
+export 'union.dart';
+export 'error.dart';