+++ lib/union.dart
@@ -1,0 +1,1 @@
+export 'src/union/union_type.dart';