+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # CHANGELOG type_utils
 
+## 0.2.0
+
+- Added Union Base Classes
+- Added Error Base Classes
+
 ## 0.1.1
 
 - Added `Builder` interface